requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
sqlparse==0.5.5
djangorestframework-simplejwt